Espero que esta vez lo aprueben lpm

+++ TP1/Ejercicio6/script.sh
@@ -50,57 +50,89 @@ listar(){ #listo
         echo "Nombre        |  Direccion de Origen"
         echo "-----------------------------------------"
         echo "$var"
+        echo "-----------------------------------------"
         #cat $logPath | awk 'BEGIN{FS="  ";OFS=" "} {print $1,$2}'
     fi
 
 }
 
-vaciar(){ #listo
-    if [[ `cat $logPath | cut -d " " -f -2` = "" ]]; then
-        echo "La papelera ya esta vacia"
-        return;
+
+getNewName(){
+    #if i do have a second argument
+    nom="$1"
+    num="$2"
+    if [[ $2 != "" ]]; then
+        awk -v var="$nom" 'BEGIN{FS="  ";OFS="  "}{if(var == $1) print}' $logPath | awk -v var=$num 'BEGIN{FS="  ";OFS="  "}{if(NR == var) printf "%s|%d" , $1, $3}'
+    else
+        awk -v var="$nom" 'BEGIN{FS="  ";OFS="  "}{if(var == $1) print}' $logPath | awk 'BEGIN{FS="  ";OFS="  "} {printf "%s|%d" , $1, $3}'
     fi
-    rm -r $papeleraPath
-    mkdir $papeleraPath
-    > $logPath
-    echo "Papelera vaciada"
+}
+
+getID(){
+    nom="$1"
+    num="$2"
+    if [[ $2 != "" ]]; then
+        awk -v var="$nom" 'BEGIN{FS="  ";OFS="  "}{if(var == $1) print}' $logPath | awk -v var=$num 'BEGIN{FS="  ";OFS=" "}{if(NR == var) printf "%d" , $3}'
+    else
+        awk -v var="$nom" 'BEGIN{FS="  ";OFS="  "}{if(var == $1) print}' $logPath | awk 'BEGIN{FS="  ";OFS=" "} {printf "%d" , $3}'
+    fi
+}
+
+getAppearances(){
+    awk 'BEGIN{FS="  "; OFS="  "}{print $1}' $logPath | awk -v var="$1" 'BEGIN{FS="  ";OFS="  "}{if(var == $1) print var}' | awk 'END{print NR}'
+}
+
+getPath(){
+    if [[ $2 != "" ]]; then
+        awk -v var="$1" 'BEGIN{FS="  ";OFS="  "}{if(var == $1) print}' $logPath | awk -v var=$2 'BEGIN{FS="  ";OFS="  "}{if(NR == var) printf "%s" , $2}'
+    else
+        awk -v var="$1" 'BEGIN{FS="  ";OFS="  "}{if(var == $1) print}' $logPath | awk 'BEGIN{FS="  ";OFS="  "} {printf "%s" , $2}'
+    fi
+    
+}
+
+showOptions(){
+    awk -v var="$1" 'BEGIN{FS="  ";OFS="  "}{if(var == $1) print}' $logPath | awk 'BEGIN{FS="  ";OFS="  "} {printf "%d - %s %s\n" , NR, $1, $2}'
 }
 
 recuperar(){ #listo
 
-    if [[ `cat $logPath | grep -w "$1" | awk 'END{print NR}'` -gt 1 ]]; then
-        grep -w "$1" < $logPath | awk 'BEGIN{FS="  ";OFS=" "}{printf "%d - %s %s\n", NR, $1, $2}'
+    app=$(getAppearances "$1")
+    
+
+    if [[ $app -gt 1 ]]; then
+        showOptions "$1"
         echo "Cual queres recuperar?"
         read archNum
-        nuevoNombre="`grep -w "$1" < $logPath | awk -v var=$archNum 'BEGIN{FS="  ";OFS=" "}{if(NR == var) printf "%s|%d" , $1, $3}'`"
-        IDarch=`grep -w "$1" < $logPath | awk -v var=$archNum 'BEGIN{FS="  ";OFS=" "}{if(NR == var) print $3}'`
+        nuevoNombre=$(getNewName "$1" $archNum)
+        IDarch=$(getID "$1" $archNum)
         mv $papeleraPath/"$nuevoNombre" $papeleraPath/"$1" 
 
         #si es un file
         if [[ -f $papeleraPath/"$1" ]]; then
-            cp $papeleraPath/"$1" "`grep -w "$1" < $logPath | awk -v var=$archNum 'BEGIN{FS="  ";OFS=" "}{if(NR == var) print $2}'`"
+            cp $papeleraPath/"$1" "$(getPath "$1" $archNum)"
             rm $papeleraPath/"$1" 
         #si es un dir"
         elif [[ -d $papeleraPath/"$1" ]]; then 
-            cp -r $papeleraPath/"$1" "`grep -w "$1" < $logPath | awk -v var=$archNum 'BEGIN{FS="  ";OFS=" "}{if(NR == var) print $2}'`" 
+            cp -r $papeleraPath/"$1" "$(getPath "$1" $archNum)"
             rm -r $papeleraPath/"$1" 
         fi
 
         sed -i "/$IDarch/d" $logPath
 
 
-    elif [[ `cat $logPath | grep -w "$1" | awk 'END{print NR}'` -eq 1 ]]; then
-        nuevoNombre="`grep -w "$1" < $logPath | awk 'BEGIN{FS="  ";OFS=" "}{printf "%s|%d" , $1, $3}'`" #consigo el nuevo nombre del archivo dentro de la papelera
+    elif [[ $app -eq 1 ]]; then
+        nuevoNombre=$(getNewName "$1") #consigo el nuevo nombre del archivo dentro de la papelera
         mv $papeleraPath/"$nuevoNombre" $papeleraPath/"$1" #le cambio el nombre al original
-        IDarch=`grep -w "$1" < $logPath | awk 'BEGIN{FS="  ";OFS=" "}{print $3}'`
+        IDarch=$(getID "$1") 
 
         #si es un file
         if [[ -f $papeleraPath/"$1" ]]; then
-            cp $papeleraPath/"$1" "`grep -w "$1" < $logPath | awk 'BEGIN{FS="  ";OFS=" "}{print $2}'`" 
+            cp $papeleraPath/"$1" "$(getPath "$1")" 
             rm $papeleraPath/"$1" 
         #si es un dir
         elif [[ -d $papeleraPath/"$1" ]]; then 
-            cp -r $papeleraPath/"$1" "`grep -w "$1" < $logPath | awk 'BEGIN{FS="  ";OFS=" "}{print $2}'`" 
+            cp -r $papeleraPath/"$1" "$(getPath "$1")" 
             rm -r $papeleraPath/"$1"
         fi
 
@@ -130,12 +162,14 @@ eliminar(){ #listo
 }
 
 borrar(){
-    if [[ `cat $logPath | grep -w "$1" | awk 'END{print NR}'` -gt 1 ]]; then
-        grep -w "$1" < $logPath | awk 'BEGIN{FS="  ";OFS=" "}{printf "%d - %s %s\n", NR, $1, $2}'
+    app=$(getAppearances "$1")
+
+    if [[ $app -gt 1 ]]; then
+        showOptions "$1"
         echo "Cual queres borrar?"
         read archNum
-        nuevoNombre=`cat $logPath | grep $1 | awk -v var=$archNum 'BEGIN{FS="  ";OFS=" "}{if(NR == var) printf "%s|%d" , $1, $3}'` 
-        IDarch=`grep -w "$1" < $logPath | awk -v var=$archNum 'BEGIN{FS="  ";OFS=" "}{if(NR == var) print $3}'`
+        nuevoNombre=$(getNewName "$1" $archNum)
+        IDarch=$(getID "$1" $archNum)
         mv $papeleraPath/"$nuevoNombre" $papeleraPath/"$1" 
 
         #si es un file
@@ -149,10 +183,10 @@ borrar(){
         sed -i "/$IDarch/d" $logPath
 
 
-    elif [[ `cat $logPath | grep -w "$1" | awk 'END{print NR}'` -eq 1 ]]; then
-        nuevoNombre=`cat $logPath | grep -w "$1" | awk 'BEGIN{FS="  ";OFS=" "}{printf "%s|%d" , $1, $3}'` 
+    elif [[ $app -eq 1 ]]; then
+        nuevoNombre=$(getNewName "$1") 
         mv $papeleraPath/"$nuevoNombre" $papeleraPath/"$1" 
-        IDarch=`grep -w "$1" < $logPath | awk 'BEGIN{FS="  ";OFS=" "}{print $3}'`
+        IDarch=$(getID "$1")
 
         #si es un file
         if [[ -f $papeleraPath/"$1" ]]; then
@@ -167,6 +201,16 @@ borrar(){
         echo "El archivo $1 no se encuentra en la papelera de reciclaje"
         return 1
     fi
+}
+vaciar(){ #listo
+    if [[ `cat $logPath | cut -d " " -f -2` = "" ]]; then
+        echo "La papelera ya esta vacia"
+        return;
+    fi
+    rm -r $papeleraPath
+    mkdir $papeleraPath
+    > $logPath
+    echo "Papelera vaciada"
 }
 
 zipHandler(){
@@ -193,14 +237,19 @@ fi
 #si no existe la papelera, la creo
 if [[ -f $papeleraZip ]]
 then
-    unzip $papeleraZip -d $papeleraPath >> /dev/null
+    unzip -q $papeleraZip -d $papeleraPath >> /dev/null
     if [[ -f $logPath ]]; then
         echo "La papelera ya existe" >> script.log
     else
         echo "La papelera existente no fue creada con este script"
-        rm -r $papeleraPath
-
+        rm $papeleraZip
+        if [[ -d $papeleraPath ]]; then
+            rm -r $papeleraPath
+        fi
         mkdir $papeleraPath
+        touch $logPath
+        zipHandler
+        exit 1
     fi
 else
     mkdir $papeleraPath
@@ -238,7 +287,7 @@ elif [[ $1 = "--recuperar" ]]; then #param: archivo
     if [[ $ruta != "" ]]; then
         recuperar "$ruta"
         if [[ $? -eq 0 ]]; then
-            echo "El archivo $ruta se ha recuperado" 
+            echo "El archivo "$ruta" se ha recuperado" 
         else
             zipHandler
             exit 1
@@ -267,9 +316,9 @@ elif [[ $1 = "--eliminar" ]]; then #param: archivo
     fi
 elif [[ $1 = "--borrar" ]]; then #param: archivo
     if [[ $ruta != "" ]]; then
-        borrar $ruta
+        borrar "$ruta"
         if [[ $? -eq 0 ]]; then
-            echo "El archivo $ruta se ha borrado de la papelera de reciclaje" 
+            echo "El archivo "$ruta" se ha borrado de la papelera de reciclaje" 
         else
             zipHandler
             exit 1;
@@ -284,3 +333,6 @@ elif [[ $1 = "--borrar" ]]; then #param: archivo
 fi
 
 zipHandler
+
+
+#TODO: replicar lo de recuperar en el resto de recuperar y en borrar (mejor si es en una funcion aparte)
